ref(api): split validator tests
